requirement.txt
asgiref==3.4.1
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.2.0
django-excel==0.0.10
django-import-export==2.6.1
django-js-asset==1.2.2
django-simple-history==3.0.0
et-xmlfile==1.1.0
Faker==9.9.0
filelock==3.4.0
gunicorn==20.1.0
lml==0.1.0
MarkupPy==1.14
numpy==1.21.4
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
platformdirs==2.4.0
pycodestyle==2.8.0
pyexcel==0.6.7
pyexcel-io==0.6.5
pyexcel-webio==0.1.4
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
text-unidecode==1.3
texttable==1.6.4
toml==0.10.2
virtualenv==20.10.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
